Refactor into modular files and update build docs
